Fix test that's breaking pre-release workflow
We had to adjust E2E tests when we removed H2 support.
That was done in #33212.

But this broke the pre-release workflow.
That workflow doesn't have necessary docker services needed for this test
(postgres and mysql).

On top of that, adding both databases in this test is kind of redundant,
and definitely an overkill. We just want to make sure we can add a database.
So, why not use SQLite that doesn't require external services?

This PR accomplishes exactly that.
It simplifies and speeds up this particular test by using SQLite database.