Backport: Fix e2e skipped test naming 2 (#33804) (#33823)
The original attempt was #33777.

As revealed in #33799, the issue hasn't been fully fixed.
Take a look at this run: https://github.com/metabase/metabase/actions/runs/6110515399

The skipped test names are now in this format: e2e-tests-collectionsslow-ee
Let's see how do we create a name:

name: e2e-tests-${{ matrix.folder }}${{ matrix.context }}-${{ matrix.edition }}
So it's obvious that both folder and context are included, when in reality we need only one of the two.

The flaw in the logic was trying to remove the runner from the matrix. At least ONE matrix parameter needs to be different in the slow context. We achieve this by setting the runner to ubuntu-latest instead of ubuntu-22.04. This should now finally fix the names.

[ci skip]