Make swap functions respect metric id
When aggregation was to be transformed to field, and that field
was comming from deeper level nested metric, id from deeper metric was
used errorneosly. In case that current aggregation is from metric.
Now id is of "this aggregation" is swapped for id of deeper aggregation
and can correctly be provided upwards.

Code still in debug mode.