Pipeline seam without changing behavior
we create a notion of "interesting", but it's 100% derived from the
groupings in the card templates in the dash template. We then match when
"interesting"s are satisfied with bound dimensions and then in make card
we can use that info.

In the future, this means that we can get affinities from
anywhere (LLMs, extra stuff, or just `(:affinities dashboard-template)`,
but for the meantime we aren't changing behavior, just making it easier
to change behavior in the future.

It still names the affinity based on the card name. So in this step i'm
just (attempting) to filter out cards from the dashboard to ones we
think it should find interesting.

However, five tests are failing.

```clojure
core-test=> (clojure.test/run-test model-with-joins-test)
<lots of junk>
Ran 1 tests containing 26 assertions.
3 failures, 0 errors.
{:test 1, :pass 23, :fail 3, :error 0, :type :summary}
```

Looking at the cards from the template that "we-identified" with
interesting, and the cards "it" identifies if you pass the whole
dashboard template through and you get the following. I don't know why
it finds so many more cards than my way. Need to investigate.

```clojure
{:we-identified ["CountByState"
                 "RowcountLast30Days"
                 "NumberDistribution"
                 "DistinctFKCounts"
                 "Rowcount"]
 :it ["Rowcount"
      "RowcountLast30Days"
      "NumberDistribution"
      "CountByState"
      "CountByCreateDate"
      "NumberOverCreateDate"
      "DayOfWeekCreateDate"
      "HourOfDayCreateDate"
      "DayOfMonthCreateDate"
      "MonthOfYearCreateDate"
      "QuerterOfYearCreateDate"]}
```