using affinities
had to update the affinities map to handle multiple definitions of the
same affinity name.

```clojure
core=> (let [affinities (-> ["table" "GenericTable"]
                            dashboard-templates/get-dashboard-template
                            dash-template->affinities-map)]
         (affinities "HourOfDayCreateDate"))
[["HourOfDayCreateDate"
  {:dimensions ["CreateTimestamp"], :metrics ["Count"], :score 50}]
 ["HourOfDayCreateDate"
  {:dimensions ["CreateTime"], :metrics ["Count"], :score 50}]]
```

Here there were two cards defined as "HourOfDateCreateDate". One looked
for a createtimestamp, the other for a createtime. When treating the
affinities as unique names, the one with createtime clobbered the
createtimestamp and then it went unsatisfied.

Now we keep both definitions around. The group by is stable so the
timestamp comes first and in match that one will be matched.