Updating dash-template->affinities-map to dash-template->affinities.
Two key changes were made:
- The card name is now a part of the 'affinity' object (under the `:affinity-name` key).
  This flattening should also make affinities easier to deal with and more flexible
  (e.g. arbitrary groupings).
- All base dimensions are exploded out into these affinity objects under the `:base-dims` key.
  Note that this may result in growth of the number of affinities when named items are
  repeated with different definitions. For example, card names are not unique, resulting in
  N affinities per card. Also, metrics and filters need not be unique. GenericTable, for
  example, has 6 definitions of the Last30Days filter. This will result in 6X the number of
  affinities created for each card using that filter.

This change encapsulates affinities, as before you'd need to know what dimensions underlied
any given metric or filter. ATM, we do not package each filter or metric definition in each
affinity object, but perhaps that would be worth doing in the future for complete encapsulation.

This change also allows for very simple matching logic. To match card affinities, for example,
you just filter all affinities for which the affinity dimensions are a subset of the provided
dimensions.