Change shape of affinities
The "satisfied-affins" are of this shape:
```clojure
{"AverageIncomeByMonth" #{"Income" "Timestamp"},
 "AverageDiscountByMonth" #{"Income" "Discount" "Timestamp"}}
```

And they are ordered so it can drive the `make-cards` function in the
future. The idea is for right now we'll look up the card based on
affinity name, and when multiple cards found, the set of dimensions they
depend on. And that will drive the layout.

But in the future, just the affinity itself will drive how we make a
card layout. This is the firs step towards that.