Fix back to x-ray dashboard flake
No need to load thousands of row and to rely on overly complicated automagic calculations.
The only thing this test wants to know whether or not we display "back to dashboard" button.

By greatly simplifying the underlying model, automagic dashboard becomes MUCH simpler.
The test will also be significantly faster as a result of this change.