Fix archived items tooltip that's missing the item type (#33970)
* Make archived item rely on the `model` for its type

The `type` prop doesn't exist in the list of items.
`models` on the other hand exists and it shows the underlying item model,
which almost 1:1 matches the desired type.

Almost, because we can questions "card" in the code.
That will be accounted for in the next commit.

* Handle `card` and `dataset` models

Show "question` and "model" in the UI instead.

* Move `getTranslatedEntityName` to the common utils

* Use translated entities for the archived item tooltip

* Switch back to the implicit return

* Lowercase the translated model name

* Add missing string translation

* Add unit test for `getTranslatedEntityName`