Auditv2 dir or zip ia loading (#34064)
* improve loading behavior for ia content

first, look for the zip file. this file will be here when running as an
uberjar. when found, unzip it to plugins/instance_analytics

If that's not found, look for the resources/instance_analytics dir,
which will be copied to plugins/instance_analytics

The serialization piece doesn't care where the yaml data in
plugins/instance_analytics came from, it just loads it as usual.

* remove redundant comments

* linter

* alter test to ignore both analytics content locations

* only load IA content when a source for it is found