Fix the flaky test 'downgrade-dashboard-tabs-test'  (#34115)
* Fix the flaky test 'downgrade-dashboard-tabs-test' by bumping the
migration version we upgrade to during the test. I'm not fully sure why
it was flaky in the first place. For some reasons the rollback step
doesn't get triggered on MySQL even though liqubiase does acknowledge
that the migration got ran?? eitherway bumping the version seems to take
the flaky a way.