Add `:dataset` to the `POST /api/card` endpoint
Depending on this key's value, we might either create a model when saving a card, or not.

It already exists in some tests and in the `create-card!` function but it was never explicitly defined in this endpoint.
That is a shame because we generate API documentation based on these keys.