Added an endpoint GET "/_affinity_cards/:entity-type/field-ids/:field-ids" that will generate cards that match the affinities of the provided fields and entity type. The leading underscore in the path is because there's some weird path matching thing going on in the router. We should definitely rename this if we move forward with it.
Given a discussion that was had about our path forward, IDK if we actually
want to keep these endpoints. However, I did some additional sleuthing on
how we pass everthing in to make-cards -> card-candidates -> build-dashcard.
I think we really should focus on card-candidates (potentially with a
different name) as the entry point for a generic card producer given
affinity sets, a card definition, and supporting data. Even if this PR
is rejected, we should probably pick out these tweaks as they make things
a bit easier to use.