Added some ns-level comment blocks to show the three major blocks within the `metabase.automagic-dashboards.metric-x-rays` namespace:
1. The dashcard generation code, which dispatches on instantiated
   affinities and could be its own ns.
2. The global affinity definitions, which could be their own ns.
3. The code to instantiate affinities, which could also be its own ns,
   probably a driver ns.

Also not that the API path for this is GET "/metric/metric/metric/:metric-id",
which is totally temporary. It's tough to come up with a name that doesn't
conflict with all of the existing path param URLs.

I wonder if this code should all go into a separate package, `dynamic_dashboards`.