X-ray dashcards with semantic type of Category and base type of Boolean were creating unparseable mbql queries.
This is an example of a query generated previous to the fix:

```clojure
; Source table is Accounts, field 8 is LEGACY_PLAN
(qp/process-query
  {:type :query,
   :database 1,
   :query
   {:source-table 7,
    :breakout     [[:field 8 nil]],
    :aggregation  [["count"]],
    :order-by     [[:desc [:aggregation 0]]]}})
```
This results in the following error:

Execution error (ExceptionInfo) at metabase.util.malli.fn/validate (fn.clj:150).
Invalid output: {:stages [{:order-by [nil [nil nil ["an expression that can be compared with :> or :<" "an expression that can be compared with :> or :<"]]], :source-card ["missing required key"]}]}

The fix is simply to ensure that the base type of the field is present in the field
attributes if the field is of the form `[:field id attributes]`.

Here is the fixed query, which works:

```clojure
(qp/process-query
  {:type :query,
   :database 1,
   :query
   {:source-table 7,
    :breakout     [[:field 8 {:base-type :type/Boolean}]],
    :aggregation  [["count"]],
    :order-by     [[:desc [:aggregation 0]]]}})
```