This is a small set of changes pulled from another PR whose contents (#34228)
probably won't be merged. The intent of these changes is to be
explicit in what is passed into each of the card creation functions,
ensure consistency of types, and eliminate redundant or non-required
arguments.

`build-dashcard` was the target of this simplification. The
`available-values` argument is no longer necessary as that data is
sourced from other arguments. the `satisfied-*` arguments are no
longer seqs of satisfied filter, metric, and dimension definitions,
but maps of names of those items to the definitions. This is only
needed because satisfied metrics needs metrics names when adding
metadata. However, all values were modified into a consistent shape.

In other areas, if an aggregate object with several keys is passed
to a function that uses only one element of the object, only the
inner element is passed.

In card-candidates, `:matches` is dissoced from `satisfied-dimensions`
to remove the extra payload that isn't needed in that argument.
This makes it clear that satisfied field values are passed into card
creation functions via the `satisfied-bindings` and not cruft from
earlier stage.

As an aside, one potential fix that would substantially improve the
readibility of this entire codebase would be to move the element
names in the one key to value maps as is done for dimensions,
metrics, filters, and cards into the elements as a named key.
Carrying that shape forward from the yaml files requires a ton of
first this or zipmap that to recover the names into whatever the
current context is.