Added implementation of `find-metrics`. This adds grounded metrics of two types: - linked metrics such as user defined metrics (from the metrics table) that link to a table - generated metrics, which come from a template
`find-metrics` produces a sequence of metrics such as is shown here:

```clojure
[
{:metric-name "Churn",
  :metric-definition {:source-table 7,
                      :aggregation [[:count]],
                      :filter [:< [:datetime-diff [:field 3 nil] [:field 13 nil] :month] 3]},
  :metric-interestingness :???}
 {:metric-name "Count", :metric-title nil, :metric-interestingness 0, :metric-definition {:aggregation ["count"]}}
 {:metric-name "MinJoinDate",
  :metric-title "Earliest Joint Date",
  :metric-interestingness 5,
  :metric-definition {:aggregation ["min" [:field 13 nil]]}}
]
```

Note that metric interestingness is derived from the semantic types of the fields used in the metric so could be dropped.

Also, to ground fields to types in the metric definition, we must look up potential bindings in the db. Do we want to do this at an earlier stage? Doing this earlier may reduce db hits, but it might also lead towards generation of a larger, messier, context object.