Added implementation of `make-combinations`, which takes a seq of grounded metrics and expands that seq to a (generally) larger seq of grounded metrics with dimensions added. These dimensions will generally become breakouts of some sort downstream when the final query is constructed.
Q: Should the query actually be updated in the make-combinations
function? I'm actually thinking maybe yes. This would give a nice
pipeline in which every element in the sequences we are creating
are unique queries that can then be applied as desired -- cards,
reports, etc.

The implementation itself is a mapcat over `make-metric-combinations`,
which determines which of the ground dimensions are already used in
the metric query then checks to ensure those are contained in the
affinitiy set. Any remaining affities must then be bound to possible
dimensions. Once this is done the results are returned.

An example of the full invocation looks like this:
```clojure
(let [template-name          "GenericTable"
      entity                 (t2/select-one :model/Table :name "ACCOUNTS")
      template               (dashboard-templates/get-dashboard-template ["table" template-name])
      {template-dimensions :dimensions
       template-metrics    :metrics} template
      semantic-affinity-sets (keys (magic/semantic-affinities template))
      base-context           (#'magic/make-base-context (magic/->root entity))
      ground-dimensions      (find-dimensions base-context template-dimensions)
      metric-templates       (normalize-metrics template-metrics)
      grounded-metrics       (concat
                               (grounded-metrics metric-templates ground-dimensions)
                               (linked-metrics entity))]
  (->> (make-combinations
         grounded-metrics
         ground-dimensions
         semantic-affinity-sets)
       (mapv (fn [ground-metric-with-dimensions]
               (-> ground-metric-with-dimensions
                   (update :grounded-metric-fields (partial map (juxt :id :name)))
                   (update :dimensions #(update-vals % (juxt :id :name))))))))
```