Added the function `generate-dashboard` to `metabase.automagic-dashboards.core` that primarily uses functions from the foo ns.
This function is basically a pipeline that:
- grounds dimensions
- grounds metrics
- expands the grounded metrics to include dimensions in their affinity sets
- expands those expanded metrics to generate cards
- does a layout and produces a dashboard

The beauty of this is that at every stage you are simply mapping from one
sequence of intermediate values to another. Each seq may expand or contract
based on the satisfiabily of matches, but each stage is fully inspectable and
you can pull an element off of it and attempt the next stage on just that
element as data.

One consideration for change is the card layer. `ground-metric->card` adds
the dataset query then creates the card. The next stage, `card->dashcard`
takes those cards and creates a dashcard. I think it would be a very good idea
to split `ground-metric->card` such that it fully defines the `dashcard_query`
in one stage _then_ creates cards downstream. The current challenge with this
is that breakout refinements (e.g. how you aggregate a time into day of week,
month of year, etc.) are baked into the card definitions from the card templates.
I think this should be broken out as we refine and add the query refinements to
the affinity sets and NOT have them be part of the card definition.

This PR also adds a rudimentary layout function that seems to go dumb not matter
what I try (a bunch of stuff goes down to the right no matter how griddish the
layout is). IDK

Here is the current `generate-dashboard` driver function that uses the
pipeline described above:

```clojure
(defn generate-dashboard
  "TODO - DOCS"
  [entity {template-dimensions :dimensions
           template-metrics    :metrics
           template-cards      :cards
           :as                 template}]
  (let [{:keys [source] :as base-context} (make-base-context (->root entity))
        ground-dimensions   (foo/find-dimensions base-context template-dimensions)
        affinities          (dash-template->affinities template ground-dimensions)
        affinity-set->cards (foo/make-affinity-set->cards ground-dimensions template-cards affinities)
        metric-templates    (foo/normalize-metrics template-metrics)
        grounded-metrics    (concat
                              (foo/grounded-metrics metric-templates ground-dimensions)
                              (foo/linked-metrics entity))]
    (->> grounded-metrics
         (foo/make-combinations ground-dimensions (map :affinity-set affinities))
         (foo/ground-metrics->cards source affinity-set->cards)
         (map foo/card->dashcard)
         foo/do-layout
         (foo/create-dashboard {:dashboard-name "FOO"}))))
```