Moved linked-metrics into metabase.automagic-dashboards.core. It is now called as part of ->root and results are added under the `:linked-metrics` key. Added docstrings to all new functions.
The primary driver for this new API is generate-dashboard, in
metabase.automagic-dashboards.core, which has nearly the same
shape as what we'd stubbed out as apply-dashboard-template-refactor.

One item we might want to address is if we should split up the new
generate-dashboard into two functions, one that has no knowledge of
dashboards and only creates grounded entities and another that
creates dashcards, does the layout, and then creates the dashboard.