update the team label names in .github/team.json
Remove the emojis from the team labels. With the emojis, the "auto-add to project" github automated workflow doesn't work, and we rely on it for the #bug-triage channel on slack. After several attempts of escaping the emojis, they've been removed from the labels, and now have to be removed here as well for the auto-assignment of PRs to continue working.