Add Kondo rule to prevent dynamic vars from taking over
See also: #34030

Rather than undertake a huge refactor right now, I thought it'd be wiser
to add a Kondo rule to prevent the problem from getting worse, then
gradually refactor out the exceptions (as Bryan started to do in #34030).

I'm open to pushback about exactly which namespaces are or are not
included, but this set seemed like a sensible place to start.