Modifying code to remove semantic dimension mapping. Now we just use metric and dimension names for matching.
The key addition to this commit is `combination/metrics+breakouts` which
is currently a misnormer. It should be `grounded-metrics->dashcards`, but
we might want to 2-step this.

To make this work, user defined metrics need *some* way to provide potential
dashcards. I've implemented this as a multimethod, but it could be a map
loaded from a template or whatever.