Remove default table_id for metrics
Metrics used to define a default

```clojure
   :model/Metric
   (fn [_] (default-timestamped
             {:creator_id  (rasta-id)
              :definition  {}
              :description "Lookin' for a blueberry"
              :name        "Toucans in the rainforest"
              :table_id    (data/id :checkins)}))
```

the problem is that you cannot make a metric on any other dataset except
test-dataset. Because computing the defaults would throw when the table
checkins was not found.

Most tests already supplied a table_id (not knowing they _had_ to be
from test-dataset and not sample-dataset, etc). Only a few places needed
a value added.

Segments also have the same issue. We can clean that up in a future PR.