Recognize user-defined metric candidates for the summary section
Seems like viz-type is a possible predictor for which cards might go
well in the summary.

```clojure
core=> (map (fn [template]
              (letfn [(body [card-def]
                        (-> card-def vals first))
                      (summary? [card-body]
                        (#{"Overview"} (:group card-body)))]
                (let [cards (interesting/normalize-seq-of-maps :card
                                                               (filter (comp summary? body)
                                                                       (:cards template)))]
                  (affinities->viz-types cards (constantly true)))))
            (dashboard-templates/get-dashboard-templates ["table"]))
({}
 {#{} #{["scalar" {}]}}
 {#{} #{["scalar" {}]},
  #{"Timestamp"} #{["line" {}] ["smartscalar" {}]}}
 {#{"Day"} #{["line" {}]}}
 {#{} #{["scalar" {}]},
  #{"Timestamp"} #{["line" {}] ["smartscalar" {}]}}
 {#{} #{["scalar" {}]},
  #{"JoinDate"} #{["line" {}] ["smartscalar" {}]}})
```

I'm getting the affinities->viz-types for the summary cards of all the
templates. We could make this smarter to match on the dim-set as well if
we like (would allow for line graphs to appear).

But the idea is find one and only one card with that viz type and throw
it in the summary ("Overview") section.