Fix metric x-rays
two issues:
1. we were getting linked metrics to a metric. this doesn't make
sense. Metrics are linked to a table. Metrics on the same table are
not (necessarily) linked.
2. This hid the fact that we weren't injecting "this" as a metric when
doing the template. A reminder what metric templates look like:

```yaml
  - ByState:
      group: Geographical
      title: "[[this]] per state"
      metrics: this
      dimensions: State
      visualization:
        map:
          map.type: region
          map.region: us_states
```

It expects to match `metrics: this` to the metric you are x-raying. So
we inject a metric named "this" to the bound metrics and it all works™