Moved `filter-referenced-fields` above `generate-dashboard` because it needs to be loaded before it is used in the `:param_fields` clause.
Added `add-field-self-reference` to `interesting/identify` as a
necessary but not sufficient condition to fix the current tests.

Note that prior to the above change we would not bind to cards such as
field/Number/Summary statistics that have 5 metrics (not due to the number
of metrics, but because they were self-referential). Now we get too many
cards because the for comprehension in `grounded-metrics->dashcards` is
unrolling the card metrics as `card-metric-name      card-metrics`
when it needs to combine those per card. My first attempt to do this
and combine the aggregates (which seemed to make sense) caused a huge
number of spec errors. Gonna baby step this.