Refactored `generate-dashboard` into two functions `generate-base-dashboard` and the original `generate-dashboard` itself. This is done to support the need for a partial application of the function in `indepth`. This allows us to remove the final reference to `apply-dashboard-template`.
One issue with this change is that both invocations of the split function
require the ground values (dimensions, metrics, and filters). To make this
work, I moved the computation of ground filters into interesting/find-values
and then passed raised the full map of all three of these items outside of
`generate-dashboard` and `generate-base-dashboard`. This actually looks
right once it's done. It allows the generation of base building blocks, then
passes those into the machinery that works with them.

Note that we still have one failing test: `related-card-generation-test`.
Rather than try to fix this and create a larger mess, I wanted to make an
incremental commit of what I know is needed then fix the test.