Do not set `updated_at` when pinned
- add an additional `before-update` hook that removes the `updated_at` field when it and `collection_position` are the
only updated fields

- use `methodical/prefer-method!` to sequence the latter hook after the `timestamped?` hook

The biggest problem here is a really odd one: this seems to work in the database, but somehow the frontend still
reports that the card was updated. I am still figuring out why this is the case, but it may be that this requires
frontend work that I'm not qualified to do.

Fixing the backend might still be worth it, since it's probably necessary for whatever frontend fix is necessary as
well - but maybe it's not, since we seem to be ignoring `updated_at` entirely on the frontend - presumably it's cached
somewhere and would be populated eventually?